+++ src/py_account.cc
@@ -44,7 +44,7 @@ namespace {
 
   long accounts_len(account_t& account)
   {
-    return account.accounts.size();
+    return static_cast<long>(account.accounts.size());
   }
 
   account_t& accounts_getitem(account_t& account, long i)
@@ -53,7 +53,7 @@ namespace {
     static account_t * last_account = NULL;
     static accounts_map::iterator elem;
 
-    long len = account.accounts.size();
+    long len = static_cast<long>(account.accounts.size());
 
     if (labs(i) >= len) {
       PyErr_SetString(PyExc_IndexError, _("Index out of range"));
